Refactor usage of TumCabeVerification in TumCabeClient
Use a dedicated interceptor to add the verification to all requests that require it.

+++ app/src/main/java/de/tum/in/tumcampusapp/component/ui/openinghour/LocationImportAction.kt
@@ -2,11 +2,10 @@ package de.tum.`in`.tumcampusapp.component.ui.openinghour
 
 import android.content.Context
 import de.tum.`in`.tumcampusapp.R
-import de.tum.`in`.tumcampusapp.api.app.TUMCabeClient
+import de.tum.`in`.tumcampusapp.api.app.TumCabeClient
 import de.tum.`in`.tumcampusapp.api.tumonline.CacheControl
 import de.tum.`in`.tumcampusapp.database.TcaDb
 import de.tum.`in`.tumcampusapp.service.DownloadWorker
-import java.io.IOException
 import javax.inject.Inject
 
 /**
@@ -15,14 +14,15 @@ import javax.inject.Inject
 class LocationImportAction @Inject constructor(
         private val context: Context,
         private val database: TcaDb,
-        private val tumCabeClient: TUMCabeClient
+        private val tumCabeClient: TumCabeClient
 ): DownloadWorker.Action {
 
-    @Throws(IOException::class)
     override fun execute(cacheBehaviour: CacheControl) {
         val openingHours = tumCabeClient.fetchOpeningHours(context.getString(R.string.language))
-        database.locationDao().removeCache()
-        database.locationDao().replaceInto(openingHours)
+        if (openingHours != null) {
+            database.locationDao().removeCache()
+            database.locationDao().replaceInto(openingHours)
+        }
     }
 
 }
